moved testing to entrypoint
